test: split test_session_type_visibility.py into 5 per-concern files
Split the 994-line test_session_type_visibility.py (4 classes:
TestSessionTypeVisibility, TestFeatureGateVisibility,
TestExplorerBindingVisibility, TestEvidenceReaderBindingVisibility) into
per-concern files plus an arch-test registration update:

- test_session_type_visibility_fleet.py: 11 functions covering fleet /
  FLEET_DISPATCH integration + 5 regression guards (orchestrator/skill/no-type
  hides_fleet_tools). Carries pytest.mark.feature("fleet") in pytestmark.
- test_session_type_visibility_orchestrator.py: 15 functions covering
  orchestrator / skill / food-truck / cook / transitional-bridge session
  types. No feature marker (none of its tests are fleet-specific).
- test_session_type_feature_gates.py: 3 functions covering the
  _apply_session_type_visibility sole-calling-convention contract and fleet
  tag activation. Carries pytest.mark.feature("fleet") in pytestmark. The
  redundant class-level _reset_mcp_visibility fixture is dropped (the
  directory conftest _reset_mcp_tags already covers it).
- test_session_type_explorer_binding.py: 4 functions covering terminal
  explorer (allowlist) visibility. The redundant class-level
  _reset_mcp_visibility fixture is dropped.
- test_session_type_evidence_reader.py: TestEvidenceReaderBindingVisibility
  class with 6 methods + staticmethod _set_complete_identity. No class-level
  autouse fixture (mirrors the source).

Cross-cutting updates:
- tests/arch/test_feature_markers.py: removed the obsolete
  server/test_session_type_visibility.py entry from _FLEET_CLASS_MARKERS
  (the file no longer exists), and added the two new fleet files
  (test_session_type_visibility_fleet.py, test_session_type_feature_gates.py)
  to _FLEET_CROSS_DIR_FILES — they have pytest.mark.feature("fleet") in
  pytestmark, which _pytestmark_has_feature() validates.

Mechanical lift: no production code changes, no behavior changes, no test
assertion changes. Each file carries
pytestmark = [layer("server"), medium] (with feature("fleet") for the two
fleet files).

Refs: #4610
